DataHarvester/services/data_transformation/pyproject.toml
# Software/DataHarvester/services/data_transformation/pyproject.toml

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-transformation"
version = "0.1.0"
description = "Data Transformation Service for DataHarvester"
requires-python = "==3.12.8"
dependencies = [
    "fastapi==0.111.1",
    "celery==5.4.0",
    "redis==5.0.3",
    "python-frontmatter==1.1.0",
    "graphviz==0.20.3",
    "ipython==8.31.0",
    "instructor>=1.4.0",
    "anthropic>=0.31.2",
    "openai>=1.12.0",
    "pandas==2.2.3",
    "pymongo==4.6.1",
    "motor==3.3.2",
    "beanie==1.25.0",
    "python-dotenv==1.0.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "uvicorn==0.27.1",
    "jinja2==3.1.3"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["data_transformation*"]
exclude = ["tests*"]

[tool.black]
line-length = 135
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 135

[tool.pytest.ini_options]
pythonpath = [
  "src"
]